refactor(test): break out Factory Storage tests for better organization
